@@ -1,46 +1,35 @@
 {
-  "name": "frpc-web",
-  "description": "An admin web ui for frp client.",
-  "author": "fatedier",
+  "name": "frpc-dashboard",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
-    "dev": "webpack-dev-server -d --inline --hot --env.dev",
-    "build": "rimraf dist && webpack -p --progress --hide-modules"
+    "dev": "vite",
+    "build": "run-p type-check build-only",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "element-ui": "^2.5.3",
-    "vue": "^2.5.22",
-    "vue-resource": "^1.5.1",
-    "vue-router": "^3.0.2",
-    "whatwg-fetch": "^3.0.0"
-  },
-  "engines": {
-    "node": ">=6"
+    "element-plus": "^2.5.3",
+    "vue": "^3.4.15",
+    "vue-router": "^4.2.5"
   },
   "devDependencies": {
-    "autoprefixer": "^9.4.7",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-component": "^1.1.1",
-    "babel-preset-es2015": "^6.24.1",
-    "css-loader": "^2.1.0",
-    "eslint": "^5.12.1",
-    "eslint-config-enough": "^0.3.4",
-    "eslint-loader": "^2.1.1",
-    "file-loader": "^3.0.1",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^2.24.1",
-    "less": "^3.9.0",
-    "less-loader": "^4.1.0",
-    "postcss-loader": "^3.0.0",
-    "rimraf": "^2.6.3",
-    "style-loader": "^0.23.1",
-    "url-loader": "^1.1.2",
-    "vue-loader": "^15.6.2",
-    "vue-template-compiler": "^2.5.22",
-    "webpack": "^2.7.0",
-    "webpack-cli": "^3.2.1",
-    "webpack-dev-server": "^3.1.14"
+    "@rushstack/eslint-patch": "^1.7.2",
+    "@types/node": "^18.11.12",
+    "@vitejs/plugin-vue": "^5.0.3",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "@vue/tsconfig": "^0.5.1",
+    "eslint": "^8.56.0",
+    "eslint-plugin-vue": "^9.21.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^3.2.4",
+    "typescript": "~5.3.3",
+    "unplugin-auto-import": "^0.17.5",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "^5.0.12",
+    "vue-tsc": "^1.8.27"
   }
 }
